+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/core/data/database/FoodYouDatabase.kt
@@ -77,7 +77,7 @@ import com.maksimowiczm.foodyou.core.data.model.search.SearchQueryEntity
          * Remove OpenFoodFactsPagingKeyEntity
          */
         AutoMigration(from = 9, to = 10, spec = MIGRATION_9_10::class),
-        AutoMigration(from = 10, to = 11),
+        AutoMigration(from = 10, to = 11)
         /**
          * @see [MIGRATION_11_12]
          * Fix sodium value in ProductEntity. Convert grams to milligrams.
