+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/goals/ui/screen/CaloriesApp.kt
@@ -80,7 +80,7 @@ private fun CaloriesNavHost(
                 onBack = {
                     navController.popBackStack<UpdateProduct>(inclusive = true)
                 },
-                onProductUpdate = {
+                onUpdate = {
                     navController.popBackStack<UpdateProduct>(inclusive = true)
                 }
             )
