+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ProductModule.kt
@@ -1,8 +1,8 @@
 package com.maksimowiczm.foodyou.feature.product
 
-import com.maksimowiczm.foodyou.feature.product.data.ProductRepository
-import com.maksimowiczm.foodyou.feature.product.ui.create.CreateProductViewModel
-import com.maksimowiczm.foodyou.feature.product.ui.update.UpdateProductViewModel
+import com.maksimowiczm.foodyou.feature.product.domain.ProductRepository
+import com.maksimowiczm.foodyou.feature.product.ui.create.CreateProductScreenViewModel
+import com.maksimowiczm.foodyou.feature.product.ui.update.UpdateProductScreenViewModel
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
@@ -10,6 +10,6 @@ import org.koin.dsl.module
 val productModule = module {
     factoryOf(::ProductRepository)
 
-    viewModelOf(::CreateProductViewModel)
-    viewModelOf(::UpdateProductViewModel)
+    viewModelOf(::CreateProductScreenViewModel)
+    viewModelOf(::UpdateProductScreenViewModel)
 }
