+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/create/CreateProductEvent.kt
@@ -1,0 +1,5 @@
+package com.maksimowiczm.foodyou.feature.product.ui.create
+
+internal sealed interface CreateProductEvent {
+    data class Created(val id: Long) : CreateProductEvent
+}
