--- app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/create/ProductFormEvent.kt
@@ -1,7 +1,0 @@
-package com.maksimowiczm.foodyou.feature.product.ui.create
-
-internal sealed interface ProductFormEvent {
-    data object DownloadedProductSuccessfully : ProductFormEvent
-    data object CreatingProduct : ProductFormEvent
-    data class ProductCreated(val id: Long) : ProductFormEvent
-}
