--- app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/update/ProductFormEvent.kt
@@ -1,6 +1,0 @@
-package com.maksimowiczm.foodyou.feature.product.ui.update
-
-internal sealed interface ProductFormEvent {
-    data object UpdatingProduct : ProductFormEvent
-    data class ProductUpdated(val id: Long) : ProductFormEvent
-}
