+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/update/UpdateProductEvent.kt
@@ -1,0 +1,5 @@
+package com.maksimowiczm.foodyou.feature.product.ui.update
+
+internal sealed interface UpdateProductEvent {
+    data object Updated : UpdateProductEvent
+}
