+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/update/UpdateProductScreen.kt
@@ -1,0 +1,135 @@
+package com.maksimowiczm.foodyou.feature.product.ui.update
+
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.nestedscroll.nestedScroll
+import androidx.compose.ui.unit.dp
+import com.maksimowiczm.foodyou.core.domain.model.Product
+import com.maksimowiczm.foodyou.core.ui.component.BackHandler
+import com.maksimowiczm.foodyou.feature.product.ui.ProductForm
+import com.maksimowiczm.foodyou.feature.product.ui.ProductFormState
+import com.maksimowiczm.foodyou.feature.product.ui.rememberProductFormState
+import foodyou.app.generated.resources.*
+import org.jetbrains.compose.resources.stringResource
+
+@OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
+@Composable
+internal fun UpdateProductScreen(
+    product: Product,
+    onBack: () -> Unit,
+    onUpdate: (ProductFormState) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val productFormState = rememberProductFormState(product)
+    val listState = rememberLazyListState()
+    val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior()
+
+    var showDiscardDialog by rememberSaveable { mutableStateOf(false) }
+    val handleBack = { showDiscardDialog = true }
+    BackHandler(
+        enabled = true
+    ) {
+        showDiscardDialog = true
+    }
+    if (showDiscardDialog) {
+        DiscardDialog(
+            onDismissRequest = { showDiscardDialog = false },
+            onConfirm = {
+                showDiscardDialog = false
+                onBack()
+            }
+        )
+    }
+
+    Scaffold(
+        modifier = modifier,
+        topBar = {
+            TopAppBar(
+                title = { Text(stringResource(Res.string.headline_edit_product)) },
+                navigationIcon = {
+                    IconButton(
+                        onClick = handleBack
+                    ) {
+                        Icon(
+                            imageVector = Icons.Default.Close,
+                            contentDescription = null
+                        )
+                    }
+                },
+                actions = {
+                    TextButton(
+                        onClick = { onUpdate(productFormState) },
+                        enabled = productFormState.isValid
+                    ) {
+                        Text(stringResource(Res.string.action_save))
+                    }
+                },
+                scrollBehavior = scrollBehavior
+            )
+        }
+    ) { paddingValues ->
+        LazyColumn(
+            modifier = Modifier
+                .imePadding()
+                .nestedScroll(scrollBehavior.nestedScrollConnection),
+            state = listState,
+            contentPadding = paddingValues
+        ) {
+            item {
+                ProductForm(
+                    state = productFormState,
+                    modifier = Modifier.padding(horizontal = 16.dp)
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun DiscardDialog(
+    onDismissRequest: () -> Unit,
+    onConfirm: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    AlertDialog(
+        onDismissRequest = onDismissRequest,
+        modifier = modifier,
+        confirmButton = {
+            TextButton(
+                onClick = onConfirm
+            ) {
+                Text(stringResource(Res.string.action_discard))
+            }
+        },
+        dismissButton = {
+            TextButton(
+                onClick = onDismissRequest
+            ) {
+                Text(stringResource(Res.string.action_cancel))
+            }
+        },
+        text = {
+            Text(stringResource(Res.string.question_discard_changes))
+        }
+    )
+}
